fix(e2e): isolate selector resolution from browser runtime
Playwright loads guide runner modules in Node before starting browser
tests. Importing the shared DOM barrel transitively loaded
@grafana/runtime, which accessed window during module initialization and
prevented Playwright from discovering tests.

Extract environment-neutral selector catalog and resolver cores. Keep
browser-facing adapters version-aware through Grafana runtime, while the
E2E runner uses a narrow Node-safe adapter with the existing latest
selector policy.

Add regression coverage for Node imports and version-aware browser
resolution.
