+++ src/lib/dom/grafana-selector.versioned.test.ts
@@ -2,12 +2,11 @@
  * Version-aware resolution and ambiguity handling for grafana-selector.
  *
  * Uses a synthetic versioned selector tree with the package's real
- * `resolveSelectors`, so the wiring from `config.buildInfo.version` through
- * forward resolution and the reverse index is exercised end-to-end.
+ * `resolveSelectors`, so forward resolution and reverse lookup exercise
+ * explicit target versions end-to-end.
  */
 
 import { describe, it, expect } from '@jest/globals';
-
 jest.mock('@grafana/runtime', () => ({
   config: { buildInfo: { version: 'latest' } },
 }));
@@ -41,49 +40,48 @@ jest.mock('@grafana/e2e-selectors', () => {
     },
   };
 });
-
 import { config } from '@grafana/runtime';
-import { toGrafanaSelector, findGrafanaSelectorPath } from './grafana-selector';
+import { toGrafanaSelector } from './grafana-selector';
 
-function elementWithTestId(testId: string): HTMLElement {
-  const el = document.createElement('button');
-  el.setAttribute('data-testid', testId);
-  return el;
-}
-
-describe('grafana-selector — version-aware resolution', () => {
-  it('resolves forward against the running Grafana version, not latest', () => {
+import { findGrafanaSelectorPathForVersion, toGrafanaSelectorForVersion } from './grafana-selector-core';
+describe('grafana-selector browser adapter', () => {
+  it('resolves against the running Grafana version', () => {
     config.buildInfo.version = '9.0.0';
     expect(toGrafanaSelector('components.Thing.button')).toContain("[data-testid='old thing button']");
   });
+});
+
+describe('grafana-selector — version-aware resolution', () => {
+  it('resolves forward against the running Grafana version, not latest', () => {
+    expect(toGrafanaSelectorForVersion('components.Thing.button', '9.0.0')).toContain(
+      "[data-testid='old thing button']"
+    );
+  });
 
   it('resolves forward to the newer value once the running version reaches it', () => {
-    config.buildInfo.version = '12.0.0';
-    expect(toGrafanaSelector('components.Thing.button')).toContain("[data-testid='data-testid new thing button']");
+    expect(toGrafanaSelectorForVersion('components.Thing.button', '12.0.0')).toContain(
+      "[data-testid='data-testid new thing button']"
+    );
   });
 
   it('falls back to the latest values when the version is not valid semver', () => {
-    config.buildInfo.version = '1.0';
-    expect(toGrafanaSelector('components.Thing.button')).toContain("[data-testid='data-testid new thing button']");
+    expect(toGrafanaSelectorForVersion('components.Thing.button', '1.0')).toContain(
+      "[data-testid='data-testid new thing button']"
+    );
   });
 
   it('reverse-matches the value the running version renders', () => {
-    config.buildInfo.version = '9.0.0';
-    expect(findGrafanaSelectorPath(elementWithTestId('old thing button'))).toBe('grafana:components.Thing.button');
+    expect(findGrafanaSelectorPathForVersion(['old thing button'], '9.0.0')).toBe('grafana:components.Thing.button');
   });
 
   it('does not reverse-match a value from a different version', () => {
-    config.buildInfo.version = '9.0.0';
-    expect(findGrafanaSelectorPath(elementWithTestId('data-testid new thing button'))).toBeNull();
+    expect(findGrafanaSelectorPathForVersion(['data-testid new thing button'], '9.0.0')).toBeNull();
   });
 
   it('rebuilds the reverse index when the version changes', () => {
-    config.buildInfo.version = '9.0.0';
-    expect(findGrafanaSelectorPath(elementWithTestId('old thing button'))).toBe('grafana:components.Thing.button');
-
-    config.buildInfo.version = '12.0.0';
-    expect(findGrafanaSelectorPath(elementWithTestId('old thing button'))).toBeNull();
-    expect(findGrafanaSelectorPath(elementWithTestId('data-testid new thing button'))).toBe(
+    expect(findGrafanaSelectorPathForVersion(['old thing button'], '9.0.0')).toBe('grafana:components.Thing.button');
+    expect(findGrafanaSelectorPathForVersion(['old thing button'], '12.0.0')).toBeNull();
+    expect(findGrafanaSelectorPathForVersion(['data-testid new thing button'], '12.0.0')).toBe(
       'grafana:components.Thing.button'
     );
   });
@@ -91,20 +89,17 @@ describe('grafana-selector — version-aware resolution', () => {
 
 describe('grafana-selector — ambiguity rejection', () => {
   it('returns null for a value claimed by more than one selector path', () => {
-    config.buildInfo.version = '12.0.0';
-    expect(findGrafanaSelectorPath(elementWithTestId('data-testid shared value'))).toBeNull();
+    expect(findGrafanaSelectorPathForVersion(['data-testid shared value'], '12.0.0')).toBeNull();
   });
 
   it('returns a unique template match with its parameter', () => {
-    config.buildInfo.version = '12.0.0';
-    expect(findGrafanaSelectorPath(elementWithTestId('data-testid thing row alpha'))).toBe(
+    expect(findGrafanaSelectorPathForVersion(['data-testid thing row alpha'], '12.0.0')).toBe(
       'grafana:components.Tpl.row:alpha'
     );
   });
 
   it('returns null when a value matches more than one template', () => {
-    config.buildInfo.version = '12.0.0';
-    expect(findGrafanaSelectorPath(elementWithTestId('data-testid thing row alpha end'))).toBeNull();
+    expect(findGrafanaSelectorPathForVersion(['data-testid thing row alpha end'], '12.0.0')).toBeNull();
   });
 
   it('never treats an argument-ignoring selector function as a template', () => {
@@ -112,7 +107,6 @@ describe('grafana-selector — ambiguity rejection', () => {
     // appear in a probe's output unless the function interpolates its argument,
     // so Static.label ('data-testid PARAM label') must not become a template
     // matching unrelated values of the shape 'data-testid ... label'.
-    config.buildInfo.version = '12.0.0';
-    expect(findGrafanaSelectorPath(elementWithTestId('data-testid foo label'))).toBeNull();
+    expect(findGrafanaSelectorPathForVersion(['data-testid foo label'], '12.0.0')).toBeNull();
   });
 });
