+++ tests/e2e-runner/utils/selector-resolver.ts
@@ -1,0 +1,5 @@
+import { resolveSelectorForVersion } from '../../../src/lib/dom/selector-resolver-core';
+
+export function resolveSelector(reftarget: string): string {
+  return resolveSelectorForVersion(reftarget, 'latest');
+}
